Removing 2 timer threads from each ISOTPSoftSocket
ISOTPSoftSocket needs timers to handle RX and TX timeouts, which were previously
implemented in a way similar to what is done in the threading.Timer object in
python: spawning a thread for each timer.

This PR removes the two timer threads from each ISOTPSoftSocket, and replaces
them with one global static thread, implemented in the TimeoutScheduler class.

This should make the ISOTPSoftSocket more performant and fix some
ISOTPSoftSocket tests that sometimes fail due to timeouts.